refactor: 설정화면 코디네이터 분리
